requirements.txt
actionlib==1.12.1
angles==1.9.12
asn1crypto==0.24.0
bondpy==1.8.5
camera-calibration==1.15.0
camera-calibration-parsers==1.11.13
catkin==0.7.29
catkin-pkg==0.4.24
controller-manager==0.18.4
controller-manager-msgs==0.18.4
cryptography==2.1.4
cv-bridge==1.13.0
diagnostic-analysis==1.9.7
diagnostic-common-diagnostics==1.9.7
diagnostic-updater==1.9.7
docutils==0.18.1
dynamic-reconfigure==1.6.4
gazebo-plugins==2.8.7
gazebo-ros==2.8.7
gencpp==0.6.5
geneus==2.2.6
genlisp==0.4.16
genmsg==0.5.16
gennodejs==2.0.1
genpy==0.6.16
idna==2.6
image-geometry==1.13.0
interactive-markers==1.11.5
joint-state-publisher==1.12.15
kdl-parser-py==1.13.1
keyring==10.6.0
keyrings.alt==3.0
laser-geometry==1.6.7
message-filters==1.14.13
pycrypto==2.6.1
pygobject==3.26.1
pyparsing==3.0.7
python-dateutil==2.8.2
python-qt-binding==0.4.4
pyxdg==0.25
PyYAML==6.0
qt-dotgraph==0.4.2
qt-gui==0.4.2
qt-gui-cpp==0.4.2
qt-gui-py-common==0.4.2
resource-retriever==1.12.7
rosbag==1.14.13
rosboost-cfg==1.14.9
rosclean==1.14.9
roscreate==1.14.9
rosgraph==1.14.13
roslaunch==1.14.13
roslib==1.14.9
roslint==0.11.2
roslz4==1.14.13
rosmake==1.14.9
rosmaster==1.14.13
rosmsg==1.14.13
rosnode==1.14.13
rosparam==1.14.13
rospkg==1.4.0
rospy==1.14.13
rosservice==1.14.13
rostest==1.14.13
rostopic==1.14.13
rosunit==1.14.9
roswtf==1.14.13
rqt-action==0.4.9
rqt-bag==0.5.1
rqt-bag-plugins==0.5.1
rqt-console==0.4.9
rqt-dep==0.4.9
rqt-graph==0.4.11
rqt-gui==0.5.2
rqt-gui-py==0.5.2
rqt-image-view==0.4.16
rqt-launch==0.4.8
rqt-logger-level==0.4.8
rqt-moveit==0.5.10
rqt-msg==0.4.8
rqt-nav-view==0.5.7
rqt-plot==0.4.13
rqt-pose-view==0.5.8
rqt-publisher==0.4.8
rqt-py-common==0.5.2
rqt-py-console==0.4.8
rqt-reconfigure==0.5.4
rqt-robot-dashboard==0.5.7
rqt-robot-monitor==0.5.13
rqt-robot-steering==0.5.10
rqt-runtime-monitor==0.5.7
rqt-rviz==0.7.0
rqt-service-caller==0.4.8
rqt-shell==0.4.9
rqt-srv==0.4.8
rqt-tf-tree==0.6.0
rqt-top==0.4.8
rqt-topic==0.4.11
rqt-web==0.4.8
rviz==1.13.24
SecretStorage==2.3.1
sensor-msgs==1.12.8
six==1.16.0
smach==2.0.1
smach-ros==2.0.1
smclib==1.8.5
tf==1.12.1
tf-conversions==1.12.1
tf2-geometry-msgs==0.6.5
tf2-kdl==0.6.5
tf2-py==0.6.5
tf2-ros==0.6.5
topic-tools==1.14.13
urdfdom-py==0.4.6
uuv-assistants==0.6.13
uuv-auv-control-allocator==0.6.13
uuv-control-cascaded-pid==0.6.13
uuv-thruster-manager==0.6.13
uuv-trajectory-control==0.6.13
xacro==1.13.17
